pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools_scm[toml]"]
build-backend = "setuptools.build_meta"

[project]
name = "lume-torch"
description = "LUME compatible wrapper for Torch based surrogate models."
readme = "README.md"
authors = [
  { name = "SLAC National Accelerator Laboratory"},
]
keywords = ["machine learning", "accelerator physics"]
classifiers = [
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "numpy",
    "pyyaml",
    "botorch>=0.15",
    "lume-base @ git+https://github.com/slaclab/lume-base@main"
]
dynamic = ["version"]
[tool.setuptools_scm]
version_file = "lume_torch/_version.py"

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest"
]
mlflow = [
    "mlflow"
]
docs = [
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material",
    "mkdocs-include-markdown-plugin"
]

[project.urls]
Homepage = "https://github.com/lume-science/lume-torch"
Documentation = "https://lume-science.github.io/lume-torch/"
"Bug Tracker" = "https://github.com/lume-science/lume-torch/issues"

[project.license]
file = "LICENSE"

[options]
zip_safe = false
include_package_data = true

[tool.setuptools.packages.find]
include = [ "lume_torch", ]
namespaces = false

[tool.ruff]

[tool.pytest.ini_options]
log_cli_level = "info"
log_level = "debug"
testpaths = ["tests"]
